+++ sdmx/tests/test_message.py
@@ -5,7 +5,7 @@ import pytest
 
 import sdmx
 from sdmx import message
-from sdmx.model import v21 as model
+from sdmx.model import common, v21
 
 
 @pytest.mark.parametrize(
@@ -19,9 +19,63 @@ def test_compare(cls):
     assert A.compare(B) is True
 
 
+class TestDataMessage:
+    def test_update(self, caplog):
+        dm = message.DataMessage()
+
+        # A data structure with 3 dimensions
+        dsd = v21.DataStructureDefinition()
+        dim_foo = dsd.dimensions.getdefault("FOO")
+        dsd.dimensions.getdefault("BAR")
+        dsd.dimensions.getdefault("BAZ")
+
+        # 0 data sets
+        dm.update()
+        assert None is dm.observation_dimension
+        assert re.match("No DataSet in message", caplog.messages[-1])
+        caplog.clear()
+
+        # 1 data set, 0 observations
+        ds0 = v21.DataSet(structured_by=dsd)
+        dm.data.append(ds0)
+
+        dm.update()
+        assert None is dm.observation_dimension
+        assert re.match(r"^Unable to .* 1 data set\(s\)", caplog.messages[-1])
+        caplog.clear()
+
+        # 1 data set, 1 observation, 3 dimensions at observation ('flat')
+        obs0 = v21.Observation(
+            dimension=dsd.make_key(common.Key, {"FOO": "f", "BAR": "b", "BAZ": "b"})
+        )
+        ds0.add_obs([obs0])
+
+        dm.update()
+        assert None is dm.observation_dimension
+        assert re.match(r"^More than 1 dimension at observation", caplog.messages[-1])
+        caplog.clear()
+
+        # 1 data set, 1 observation, 1 dimension at observation
+        obs0.dimension = dsd.make_key(common.Key, {"FOO": "f"})
+
+        dm.update()
+        assert dim_foo is dm.observation_dimension
+        assert 0 == len(caplog.messages)
+
+        # 2 data sets, 1 dimension at observation, but different
+        ds1 = v21.DataSet(structured_by=dsd)
+        obs1 = v21.Observation(dimension=dsd.make_key(common.Key, {"BAR": "b"}))
+        ds1.add_obs([obs1])
+        dm.data.append(ds1)
+
+        dm.update()
+        assert None is dm.observation_dimension
+        assert re.match("Multiple data sets with different observ", caplog.messages[-1])
+
+
 class TestStructureMessage:
     def test_add_contains_get(self):
-        dsd = model.DataStructureDefinition(id="foo")
+        dsd = v21.DataStructureDefinition(id="foo")
         msg = message.StructureMessage()
 
         # add() stores the object
@@ -35,7 +89,7 @@ class TestStructureMessage:
         assert dsd is msg.get("foo")
 
         # add() with an object not collected in a StructureMessage raises TypeError
-        item = model.Item(id="bar")
+        item = common.Item(id="bar")
         with pytest.raises(TypeError):
             msg.add(item)
 
@@ -44,29 +98,53 @@ class TestStructureMessage:
             item in msg
 
         # get() with two objects of the same ID raises ValueError
-        msg.add(model.DataflowDefinition(id="foo"))
+        msg.add(v21.DataflowDefinition(id="foo"))
         with pytest.raises(ValueError):
             msg.get("foo")
 
     def test_dictlike_attribute_access(self):
-        dsd = model.DataStructureDefinition(id="foo")
+        dsd = v21.DataStructureDefinition(id="foo")
         msg = message.StructureMessage()
         msg.add(dsd)
 
         # Attribute access works when added to the default, empty DictLike
         assert msg.structure.foo is dsd
 
+    def test_iter_objects(self):
+        """:meth:`.iter_objects` can be used to iterate over all objects."""
+        msg = message.StructureMessage()
+
+        # Add several objects
+        msg.add(common.Codelist(id="CL_FOO"))
+        msg.add(common.Codelist(id="CL_BAR", is_external_reference=True))
+        msg.add(common.Codelist(id="CL_BAZ"))
+
+        msg.add(common.ConceptScheme(id="CS_FOO"))
+        msg.add(common.ConceptScheme(id="CS_BAR", is_external_reference=True))
+        msg.add(common.ConceptScheme(id="CS_BAZ"))
+
+        # Method accepts external_reference arg; runs; returns an iterator
+        result = msg.iter_objects()
+
+        # All objects
+        assert 6 == len(list(result))
+
+        # All objects that are not external references are included
+        assert {"CL_FOO", "CL_BAZ", "CS_FOO", "CS_BAZ"} == set(
+            obj.id for obj in msg.iter_objects(external_reference=False)
+        )
+
     def test_objects(self):
         """:meth:`.objects` can be used to access a collection according to class."""
         msg = message.StructureMessage()
-        cl = model.Codelist(id="foo")
+        cl = common.Codelist(id="foo")
         msg.add(cl)
 
-        assert cl is msg.objects(model.Codelist)["foo"]
+        assert cl is msg.objects(common.Codelist)["foo"]
 
         with pytest.raises(TypeError):
             # TypeError is raised for classes not collected in this message
-            msg.objects(model.DataSet)
+            msg.objects(v21.DataSet)
 
 
 EXPECTED = [